requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.2.6
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
textblob==0.19.0
tqdm==4.67.1
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
